langchain-learning/requirements.txt
# LangChain learning project dependencies
# Recommended baseline for the refreshed 2026-oriented examples.

langchain>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.0
langchain-text-splitters>=1.0.0
langgraph>=1.0.0

openai>=1.75.0
python-dotenv>=1.0.1
pydantic>=2.8.0
tiktoken>=0.9.0
pypdf>=5.4.0

# Optional but useful for tracing and debugging
langsmith>=0.4.0
